requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.16
psycopg2-binary==2.9.9
python-dotenv==1.0.1
email-validator==2.1.0.post1
openai